project/requirements.txt
customtkinter==5.2.0
mysql-connector-python==8.1.0
pyinstaller==5.13.2